Order Dialog Function : Add tags and amounts
