+++ client/lib/services/decorations.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 
 const kConfirmButtonColor = Color(0xFF0E4DA4);
 const kCancelButtonColor = Color(0xFFE74C3C);
-const fieldColor = Color(0xFFEAEAEA);
+const kInkColor = Color(0xFFEAEAEA);
 
 const kBoxDecorationStyle = BoxDecoration(
-  color: fieldColor,
+  color: kInkColor,
   boxShadow: [
     BoxShadow(
       color: Colors.black12,
