+++ client/lib/widget/login_text_field.dart
@@ -36,33 +36,34 @@ class _LoginTextFieldState extends State<LoginTextField> {
   @override
   Widget build(BuildContext context) {
     return TextField(
-        style: const TextStyle(fontSize: 20),
-        controller: widget.textController,
-        obscureText: widget.isPasswordField ? _textVisible : !_textVisible,
-        keyboardType: TextInputType.text,
-        decoration: InputDecoration(
-          labelText: widget.hint,
-          hintText: widget.hint,
-          enabledBorder: kOutlineInputBorder,
-          focusedBorder: kOutlineInputBorder,
-          prefixIcon: Icon(
-            widget.icon,
-            color: Colors.black38,
-          ),
-          suffixIcon: widget.isPasswordField
-              ? IconButton(
-                  splashRadius: 1,
-                  icon: Icon(
-                    _textVisible ? Icons.visibility : Icons.visibility_off,
-                    color: Theme.of(context).primaryColorDark,
-                  ),
-                  onPressed: () {
-                    setState(() {
-                      _textVisible = !_textVisible;
-                    });
-                  },
-                )
-              : null,
-        ));
+      style: const TextStyle(fontSize: 20),
+      controller: widget.textController,
+      obscureText: widget.isPasswordField ? _textVisible : !_textVisible,
+      keyboardType: TextInputType.text,
+      decoration: InputDecoration(
+        labelText: widget.hint,
+        hintText: widget.hint,
+        enabledBorder: kOutlineInputBorder,
+        focusedBorder: kOutlineInputBorder,
+        prefixIcon: Icon(
+          widget.icon,
+          color: Colors.black38,
+        ),
+        suffixIcon: widget.isPasswordField
+            ? IconButton(
+                splashRadius: 1,
+                icon: Icon(
+                  _textVisible ? Icons.visibility : Icons.visibility_off,
+                  color: Theme.of(context).primaryColorDark,
+                ),
+                onPressed: () {
+                  setState(() {
+                    _textVisible = !_textVisible;
+                  });
+                },
+              )
+            : null,
+      ),
+    );
   }
 }
