+++ client/lib/widget/order_dialog.dart
@@ -1,7 +1,14 @@
 import 'package:client/services/decorations.dart';
-import 'package:client/widget/styling_buttons.dart';
+import 'package:client/widget/styled_buttons.dart';
+import 'package:client/widget/widget_for_order_dialog/action_row.dart';
+import 'package:client/widget/widget_for_order_dialog/num_input.dart';
+import 'package:client/widget/widget_for_order_dialog/tag_input.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'widget_for_order_dialog/label_text_container.dart';
+import 'widget_for_order_dialog/amount_row.dart';
+import 'widget_for_order_dialog/filter_image.dart';
+import 'widget_for_order_dialog/product_info.dart';
 
 class OrderDialog extends StatefulWidget {
   const OrderDialog(
@@ -12,9 +19,9 @@ class OrderDialog extends StatefulWidget {
       required this.info})
       : super(key: key);
   final String product;
-  final String info;
+  final List<String> info;
   final String price;
-  final IconData img;
+  final String img;
 
   @override
   _OrderDialogState createState() => _OrderDialogState();
@@ -22,45 +29,74 @@ class OrderDialog extends StatefulWidget {
 
 class _OrderDialogState extends State<OrderDialog> {
   List<Widget> numButtons = [];
-
   List<Widget> tagButtons = [];
+  List<Widget> editButtons = [];
 
   @override
   void initState() {
-    numButtons = List.generate(
-      10,
-      (i) => Padding(
-        padding: const EdgeInsets.fromLTRB(0, 0, 10, 0),
-        child: ActionButton(
-          action: " $i",
-          color: primaryTextColor,
+    editButtons = [
+      ActionButton(
+          color: kConfirmButtonColor,
+          action: '輸入下列',
           onPress: () {
             setState(() {
-              labels = labels + "*$i \n";
+              if(labels.last.contains(RegExp(r'([*][0-9])'))){
+                labels.add('');
+              }else{
+                labels.last = '';
+                const snackBar = SnackBar(
+                  content: Text('不符合規範，因此刪除'),
+                );
+                ScaffoldMessenger.of(context).showSnackBar(snackBar);
+              }
             });
-          },
-        ),
+          }),
+      ActionButton(
+          color: kCancelButtonColor,
+          action: '清除單行',
+          onPress: () {
+            setState(() {
+              (labels.length == 1)?labels = ['']:labels.removeAt(labels.length - 1);
+            });
+          }),
+      const SizedBox(height: 48),
+      ActionButton(
+          color: kCancelButtonColor,
+          action: '清空所有',
+          onPress: () {
+            setState(() {
+              labels = [''];
+            });
+          }),
+    ];
+    numButtons = List.generate(
+      10,
+      (i) => ActionButton(
+        action: " $i",
+        color: primaryTextColor,
+        onPress: () {
+          setState(() {
+            labels[labels.length - 1] = labels[labels.length - 1] + "$i";
+          });
+        },
       ),
     );
     tagButtons = List.generate(
-      10,
-      (i) => Padding(
-        padding: const EdgeInsets.fromLTRB(0, 0, 10, 0),
-        child: ActionButton(
-          action: " tag $i",
-          color: primaryTextColor,
-          onPress: () {
-            setState(() {
-              labels = labels + "tag:$i";
-            });
-          },
-        ),
+      widget.info.length,
+      (i) => ActionButton(
+        action: widget.info[i],
+        color: primaryTextColor,
+        onPress: () {
+          setState(() {
+            labels[labels.length - 1] = widget.info[i] + '*';
+          });
+        },
       ),
     );
     super.initState();
   }
 
-  String labels = '';
+  List<String> labels = [''];
 
   @override
   Widget build(BuildContext context) {
@@ -72,22 +108,28 @@ class _OrderDialogState extends State<OrderDialog> {
           width: 900.0,
           child: Column(
             children: <Widget>[
-              Stack(children: [
-                const FilteredImage(
-                    image: AssetImage("assets/img/test_img.jpg")),
-                ProductInfo(widget: widget),
-              ]),
+              Stack(
+                  clipBehavior: Clip.none,
+                  alignment: Alignment.center,
+                  children: [
+                    FilteredImage(
+                      image: widget.img,
+                    ),
+                    ProductInfo(widget: widget),
+                  ]),
               const AmountRow(),
-              const Divider(),
-              Row(
-                children: [
-                  LabelTextContainer(labels: labels),
-                  NumInput(numButtons: numButtons),
-                  const SizedBox(
-                    width: 25,
-                  ),
-                  NumInput(numButtons: tagButtons),
-                ],
+              Padding(
+                padding: const EdgeInsets.fromLTRB(20, 10, 20, 10),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceAround,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    LabelTextContainer(labels: labels),
+                    TagsInput(tagButtons: tagButtons),
+                    NumInput(numButtons: numButtons),
+                    EditColumnForBox(buttons: editButtons),
+                  ],
+                ),
               ),
               const ActionRow(),
             ],
@@ -98,296 +140,17 @@ class _OrderDialogState extends State<OrderDialog> {
   }
 }
 
-class NumInput extends StatelessWidget {
-  const NumInput({
+class EditColumnForBox extends StatelessWidget {
+  const EditColumnForBox({
+    required this.buttons,
     Key? key,
-    required this.numButtons,
   }) : super(key: key);
-
-  final List<Widget> numButtons;
+  final List<Widget> buttons;
 
   @override
   Widget build(BuildContext context) {
     return Column(
-      children: [
-        Row(
-          children: numButtons.sublist(7),
-        ),
-        Row(
-          children: numButtons.sublist(4, 7),
-        ),
-        Row(
-          children: numButtons.sublist(1, 4),
-        ),
-        Row(
-          children: [numButtons[0]],
-        ),
-      ],
-    );
-  }
-}
-
-class ActionRow extends StatelessWidget {
-  const ActionRow({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(20, 10, 20, 10),
-      child: Row(
-        children: [
-          SizedBox(
-            width: 150,
-            height: 50,
-            child: ActionButton(
-              action: '確定',
-              color: kConfirmButtonColor,
-              onPress: () => Navigator.pop(context),
-            ),
-          ),
-          const Spacer(),
-          SizedBox(
-            width: 150,
-            height: 50,
-            child: ActionButton(
-              action: '取消',
-              color: kCancelButtonColor,
-              onPress: () => Navigator.pop(context),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class AmountRow extends StatefulWidget {
-  const AmountRow({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  State<AmountRow> createState() => _AmountRowState();
-}
-
-class _AmountRowState extends State<AmountRow> {
-  int amount = 0;
-
-  var textController = TextEditingController(text: '0');
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(0, 5, 0, 0),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          ActionButton(
-            color: primaryTextColor,
-            action: '+5',
-            onPress: () {
-              amount = amount + 5;
-              textController.text = amount.toString();
-            },
-          ),
-          const SizedBox(
-            width: 30,
-          ),
-          ActionButton(
-            color: primaryTextColor,
-            action: '+',
-            onPress: () {
-              amount = amount + 1;
-              textController.text = amount.toString();
-            },
-          ),
-          const SizedBox(
-            width: 30,
-          ),
-          Container(
-            decoration: BoxDecoration(
-              border: Border.all(color: primaryTextColor),
-              borderRadius: BorderRadius.circular(12),
-            ),
-            width: 150,
-            height: 40,
-            child: TextField(
-              controller: textController,
-              decoration: const InputDecoration(
-                hintText: "0",
-                border: InputBorder.none,
-              ),
-              textAlign: TextAlign.center,
-              readOnly: true,
-              enableInteractiveSelection: false,
-            ),
-          ),
-          const SizedBox(
-            width: 30,
-          ),
-          ActionButton(
-            color: primaryTextColor,
-            action: '-',
-            onPress: () {
-              amount = amount - 1;
-              if (amount < 0) {
-                amount = 0;
-              }
-              textController.text = amount.toString();
-            },
-          ),
-          const SizedBox(
-            width: 30,
-          ),
-          ActionButton(
-            color: primaryTextColor,
-            action: '-5',
-            onPress: () {
-              amount = amount - 5;
-              if (amount < 0) {
-                amount = 0;
-              }
-              textController.text = amount.toString();
-            },
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class FilteredImage extends StatelessWidget {
-  const FilteredImage({
-    Key? key,
-    required this.image,
-  }) : super(key: key);
-
-  final AssetImage image;
-
-  @override
-  Widget build(BuildContext context) {
-    return ColorFiltered(
-      colorFilter: const ColorFilter.mode(
-        Colors.grey,
-        BlendMode.modulate,
-      ),
-      child: Container(
-        height: 200,
-        width: double.infinity,
-        decoration: BoxDecoration(
-          borderRadius: const BorderRadius.only(
-            topLeft: Radius.circular(12),
-            topRight: Radius.circular(12),
-          ),
-          image: DecorationImage(
-            image: image,
-            fit: BoxFit.cover,
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class ProductInfo extends StatelessWidget {
-  const ProductInfo({
-    Key? key,
-    required this.widget,
-  }) : super(key: key);
-
-  final OrderDialog widget;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.start,
-      children: [
-        Padding(
-          padding: const EdgeInsets.fromLTRB(30, 30, 0, 0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                widget.product,
-                style: const TextStyle(color: Colors.white, fontSize: 30),
-              ),
-              const SizedBox(
-                height: 10,
-              ),
-              Text(
-                widget.price,
-                style: const TextStyle(color: Colors.white, fontSize: 30),
-              ),
-              const SizedBox(
-                height: 10,
-              ),
-              Text(
-                widget.info,
-                style: const TextStyle(color: Colors.white, fontSize: 30),
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class LabelTextContainer extends StatefulWidget {
-  const LabelTextContainer({
-    required this.labels,
-    Key? key,
-  }) : super(key: key);
-
-  final String labels;
-
-  @override
-  State<LabelTextContainer> createState() => _LabelTextContainerState();
-}
-
-class _LabelTextContainerState extends State<LabelTextContainer> {
-  final _scrollController = ScrollController();
-
-  void _scrollDown() {
-    _scrollController.jumpTo(_scrollController.position.maxScrollExtent);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    WidgetsBinding.instance!.addPostFrameCallback((_) {
-      if (_scrollController.hasClients) {
-        _scrollDown();
-      }
-    });
-    return Padding(
-      padding: const EdgeInsets.all(20.0),
-      child: SizedBox(
-        height: 200,
-        width: 350,
-        child: InputDecorator(
-          expands: true,
-          decoration: InputDecoration(
-            labelText: '客製化選項',
-            border: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(10.0),
-            ),
-          ),
-          //TODO tags adder
-          child: MediaQuery.removePadding(
-            context: context,
-            removeTop: true,
-            child: CupertinoScrollbar(
-              child: SingleChildScrollView(
-                controller: _scrollController,
-                child: Text(widget.labels),
-              ),
-            ),
-          ),
-        ),
-      ),
+      children: buttons,
     );
   }
 }
