+++ client/lib/widget/widget_for_order_dialog/product_info.dart
@@ -1,0 +1,46 @@
+import 'package:flutter/material.dart';
+import '../order_dialog.dart';
+
+class ProductInfo extends StatelessWidget {
+  const ProductInfo({
+    Key? key,
+    required this.widget,
+  }) : super(key: key);
+
+  final OrderDialog widget;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.start,
+      children: [
+        Padding(
+          padding: const EdgeInsets.fromLTRB(30, 30, 0, 0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                widget.product,
+                style: const TextStyle(color: Colors.white, fontSize: 30),
+              ),
+              const SizedBox(
+                height: 10,
+              ),
+              Text(
+                widget.price,
+                style: const TextStyle(color: Colors.white, fontSize: 30),
+              ),
+              const SizedBox(
+                height: 10,
+              ),
+              Text(
+                widget.info.toString(),
+                style: TextStyle(color: Colors.white, fontSize: 30),
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
